Cleanup, crate features
