+++ src/utils/encoder.rs
@@ -45,6 +45,31 @@ fn encode_chunk(chunk: Vec<u8>) -> [char; 4] {
     out
 }
 
-pub(crate) fn encode(input: String) -> String {
+pub(super) fn encode(input: String) -> String {
     String::from_iter(input.as_bytes().chunks(3).map(split).flat_map(encode_chunk))
+}
+
+#[cfg(test)]
+mod test {
+    use base64::{self, engine::general_purpose, Engine};
+    use rand::{
+        distributions::{Alphanumeric, DistString},
+        thread_rng,
+    };
+
+    use super::encode;
+
+    #[test]
+    fn b64() {
+        for _ in 0..100000 {
+            let mut rng = thread_rng();
+            let string = Alphanumeric.sample_string(&mut rng, 20);
+            let b64_encoded = general_purpose::STANDARD.encode(string.clone());
+            let my_encoded = encode(string);
+
+            if !my_encoded.eq(&b64_encoded) {
+                panic!("{my_encoded:?} != {b64_encoded:?}")
+            }
+        }
+    }
 }
